Added firing

+++ Source/Engine/Renderer/Renderer.cpp
@@ -21,7 +21,7 @@ namespace kiko {
 	void Renderer::EndFrame() {
 		SDL_RenderPresent(m_renderer);
 	}
-	void Renderer::SetColor(int r, int g, int b, int a) {
+	void Renderer::SetColor(uint8_t r, uint8_t g, uint8_t b, uint8_t a) {
 		SDL_SetRenderDrawColor(m_renderer, r, g, b, a);
 	}
 	void Renderer::DrawLine(int x1, int y1, int x2, int y2) {
